refactor: rename components shadowing built-in globals
Number, Error and Map shadowed the global objects of the same name
inside their modules, so those globals were unreachable there and a
reader cannot tell which one an identifier refers to.

Rename the map component file to match its new name as well. Its only
consumer imported it under an alias, which the rename makes unnecessary.

Refs: TILA-4426
