+++ frontend/apps/customer/src/components/UnitMap.tsx
@@ -7,7 +7,7 @@ type Props = {
   height?: string;
 };
 
-export const Map = ({ tprekId, height = "480px" }: Props): React.ReactElement | null => {
+export const UnitMap = ({ tprekId, height = "480px" }: Props): React.ReactElement | null => {
   const { t, i18n } = useTranslation();
   const mapUrl = `${mapUrlPrefix}${i18n.language}/embed/unit/${tprekId}`;
   return (
